create mapper folder
