+++ src/ws/service/friendship_event_updates.rs
@@ -16,7 +16,7 @@ use crate::{
 use super::{
     friendship_event_validator::validate_new_event,
     friendship_status_calculator::get_new_friendship_status,
-    mapper::update_request_as_event_payload,
+    mapper::events::update_request_as_event_payload,
     synapse_handler::{
         get_or_create_synapse_room_id, set_account_data, store_message_in_synapse_room,
         store_room_event_in_synapse_room,
