+++ src/ws/service/mapper/mod.rs
@@ -1,0 +1,2 @@
+pub mod events;
+pub mod payload_to_response;
